+++ lib/DotEnv.php
@@ -65,4 +65,56 @@ class DotEnv {
 
         return $value;
     }
+
+    public static function resolveEnvsInString(string $str) {
+
+        static $envs = null;
+
+        if (strpos($str, '${') === false) {
+            return $str;
+        }
+
+        if ($envs === null) {
+            $envs = array_merge(getenv(), $_ENV);
+        }
+
+        // Use regex only if '${' is found
+        if (preg_match_all('/\$\{([A-Za-z0-9_]+)\}/', $str, $matches)) {
+
+            if (count($matches[1]) === 1 && '${'.$matches[1][0].'}' === $str && isset($envs[$matches[1][0]])) {
+
+                $value = $envs[$matches[1][0]];
+
+                if ($value === 'null') {
+                    $value = null;
+                } elseif ($value === 'true') {
+                    $value = true;
+                } elseif ($value === 'false') {
+                    $value = false;
+                } elseif (is_numeric($value)) {
+                    return ($value + 0);
+                }
+
+                return $value;
+            }
+
+            foreach ($matches[1] as $key) {
+                if (!isset($envs[$key])) continue;
+                $str = str_replace('${'.$key.'}', $envs[$key], $str);
+            }
+        }
+
+        return $str;
+    }
+
+    public static function resolveEnvsInArray(&$array) {
+
+        foreach ($array as &$value) {
+            if (is_string($value)) {
+                $value = self::resolveEnvsInString($value);
+            } elseif (is_array($value)) {
+                self::resolveEnvsInArray($value);
+            }
+        }
+    }
 }