+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## WIP
+## 2.8.0 (2024-01-31)
 
 - Lazy loading of asset preview images
 - Update Vue to version 3.4.15